requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
python-slugify==4.0.1
requests==2.25.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
